requirements.txt
langchain>=0.3.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
langchain-community>=0.3.0,<1.0.0
langchain-text-splitters>=0.3.0,<1.0.0
langchain-google-genai>=2.0.0,<3.0.0
faiss-cpu==1.8.0
pypdf==4.2.0
streamlit==1.33.0
python-dotenv==1.0.1
tiktoken